+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Use default rules `None` in `adapt`
+
 ## [0.5.0] - 2022-01-25
 
 ### Added